sources/ricable/ottomator-agents/streambuzz-agent/requirements.txt
APScheduler==3.11.0
cachetools==5.5.1
fastapi==0.115.7
fastapi-cli==0.0.7
google-generativeai==0.8.4
openai==1.60.2
protobuf==5.29.3
pydantic==2.10.6
pydantic_ai==0.0.20
python-dotenv==1.0.1
Requests==2.32.3
supabase==2.12.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
